More package refactoring. Nothing to see here.
